% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{ai_chat}
\alias{ai_chat}
\title{ai chat template
Main layout tructure}
\usage{
ai_chat(..., title = "Chat IA - Assistente Inteligente")
}
\arguments{
\item{title}{Title ilustrated on browser}
}
\description{
ai chat template
Main layout tructure
}
