% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{ai_chat_main}
\alias{ai_chat_main}
\title{Main}
\usage{
ai_chat_main(
  texto1 = "Como posso ajudar voce hoje?",
  texto2 =
    "Sou seu assistente de IA. Posso ajudar com perguntas, escrever textos, explicar conceitos e muito mais.",
  logos = logo,
  sugest = NULL
)
}
\description{
Main
}
