% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_deepseek.R
\name{func_deepseek_request}
\alias{func_deepseek_request}
\title{function to call deepseek connection}
\usage{
func_deepseek_request(
  prompt,
  system = "You are a helpful assistant.",
  api_key = Sys.getenv("DEEPSEEK_API_KEY")
)
}
\arguments{
\item{prompt}{a list following the deepseek format message role}

\item{system}{providing a system prompt}

\item{api_key}{deepseek api key}
}
\value{
deepseek call output

structuring the messages
}
\description{
function to call deepseek connection
}
